requirements.txt
slack_sdk
requests
icalendar
google-api-python-client
google-auth
